refectored App Listen call and app code for easy testing purpose
