+++ package.json
@@ -25,6 +25,10 @@
   },
   "devDependencies": {
     "env-cmd": "^10.1.0",
-    "jest": "^29.1.2"
+    "jest": "^29.1.2",
+    "supertest": "^6.3.0"
+  },
+  "jest": {
+    "testEnvironment": "node"
   }
 }
